Update README.md
Health insurance costs are influenced by several demographic and health factors, making accurate prediction a challenging task. The goal of this project was to develop a machine learning model to predict insurance charges based on individual characteristics such as age, BMI, number of dependents, smoker status, and geographic region.
This analysis leveraged Python to:
Identify the most influential factors driving insurance charges, build a predictive model with strong performance and interpretability and highlight areas of improvement and limitations in the predictions.